Overhaul
Revamped M3 selector. Grouping seriral comm and all IO of the selector under a selector FB. Before components of the selector were spread out across the project. It was a mess.

Used to have a softIO program to map IO. Removing that to include IO under the FB that use them. Much more sane. Most revamping in the overhaul will be to change to this scheme.

Realized serial io was always conducted the same way so I made a simple wrapper fb and struct to avoid future mistakes.

More reorganization to come.

pyTMC pragma next.